Switched code to utilities
